ccom -- now with correctness!
Start our IDs at *1*, not 0.  This lets us differentiate the '0'
case... we use 0 to indicate that the value did not pass the
scalar function.  That is, 0 is like a global separator.

This gives really confusing semantics if you want '0' values to be
included in your scalar function that dictates which values should
be looked at.

Don't do that.